Move group list to its own component
