+++ app/src/main/java/com/rebalance/ui/screen/main/GroupScreen.kt
@@ -1,27 +1,18 @@
 package com.rebalance.ui.screen.main
 
 import android.content.Context
-import android.graphics.BitmapFactory
 import android.os.StrictMode
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Alignment.Companion.Center
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import com.rebalance.Preferences
 import com.rebalance.PreferencesData
@@ -32,13 +23,8 @@ import com.rebalance.backend.exceptions.ServerException
 import com.rebalance.backend.service.BackendService
 import com.rebalance.ui.component.main.BarChart
 import com.rebalance.ui.component.main.GroupSelection
+import com.rebalance.ui.component.main.GroupSpendingList
 import com.rebalance.utils.alertUser
-import compose.icons.EvaIcons
-import compose.icons.evaicons.Fill
-import compose.icons.evaicons.fill.Trash
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 @Composable
 fun GroupScreen(
@@ -89,14 +75,14 @@ fun GroupScreen(
             }
         } else { // if list tab
             // show list
-            DisplayList(
+            DisplayGroupList(
+                BackendService(preferences).getGroupList(groupId),
                 preferences,
                 groupId,
-                BackendService(preferences).getGroupList(groupId),
+                context,
                 refreshAndOpenGroup = { newGroupId ->
                     groupId = newGroupId
-                },
-                context
+                }
             )
         }
     }
@@ -260,189 +246,25 @@ private fun DisplayVisual(
 }
 
 @Composable
-private fun DisplayList(
+private fun DisplayGroupList(
+    data: List<Expense>,
     preferences: PreferencesData,
     groupId: Long,
-    data: List<Expense>,
-    refreshAndOpenGroup: (Long) -> Unit,
-    context: Context
+    context: Context,
+    refreshAndOpenGroup: (Long) -> Unit
 ) {
-    val groupCurrency =
-        if (groupId == -1L) "" else BackendService(preferences).getGroupById(groupId).getCurrency()
     Box(
         modifier = Modifier
             .fillMaxSize()
             .testTag("groupList"),
         contentAlignment = Center
     ) {
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(10.dp),
-            verticalArrangement = Arrangement.Top
-        ) {
-            val expensesByGlobalId: Map<Long?, List<Expense>> = data.groupBy { it.getGlobalId() }
-            items(
-                items = data.filter { it.getAmount() >= 0 },
-                itemContent = { item -> //TODO: apply filter previously
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(10.dp)
-                    ) {
-                        Column(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .background(Color.LightGray)
-                        ) {
-                            Row(
-                                modifier = Modifier
-                                    .padding(10.dp)
-                                    .fillMaxWidth(),
-                                horizontalArrangement = Arrangement.SpaceBetween
-                            ) {
-                                Text(
-                                    text = item.getDescription(),
-                                    fontSize = 14.sp,
-                                    modifier = Modifier
-                                        .padding(5.dp)
-                                )
-                                Text(
-                                    text = "${item.getAmount()} $groupCurrency",
-                                    fontSize = 14.sp,
-                                    color = Color.hsl(358f, 0.63f, 0.49f),
-                                    modifier = Modifier
-                                        .padding(5.dp)
-                                )
-
-                                val showDialog = remember { mutableStateOf(false) }
-
-                                IconButton(onClick = {
-                                    showDialog.value = true
-                                }) {
-                                    Icon(EvaIcons.Fill.Trash, "Delete expense")
-                                }
-                                if (showDialog.value) {
-                                    AlertDialog(
-                                        onDismissRequest = { showDialog.value = false },
-                                        title = { Text("Confirmation") },
-                                        text = { Text("Are you sure you want to delete this expense?") },
-                                        confirmButton = {
-                                            TextButton(onClick = {
-                                                BackendService(preferences).deleteExpenseByGlobalId(
-                                                    item.getGlobalId()
-                                                )
-                                                alertUser("Expense deleted!", context)
-                                                showDialog.value = false
-                                                refreshAndOpenGroup(groupId)
-                                            }) {
-                                                Text("Yes")
-                                            }
-                                        },
-                                        dismissButton = {
-                                            TextButton(onClick = { showDialog.value = false }) {
-                                                Text("No")
-                                            }
-                                        }
-                                    )
-                                }
-                            }
-                            Text(
-                                text = "Category: " + item.getCategory(),
-                                fontSize = 14.sp,
-                                modifier = Modifier
-                                    .padding(horizontal = 10.dp)
-                                    .fillMaxWidth()
-                            )
-                            Text(
-                                text = "Date: " + item.getDateStamp(),
-                                fontSize = 14.sp,
-                                modifier = Modifier
-                                    .padding(horizontal = 10.dp)
-                                    .fillMaxWidth()
-                            )
-                            val payer = item.getUser()
-                            if (payer != null) {
-                                Text(
-                                    text = "Payed by: " + payer.getUsername(),
-                                    fontSize = 14.sp,
-                                    modifier = Modifier
-                                        .padding(horizontal = 10.dp)
-                                        .fillMaxWidth()
-                                )
-                            }
-                            val toWhom = expensesByGlobalId[item.getGlobalId()]
-                            if (toWhom != null && toWhom.isNotEmpty()) {
-                                Column(
-                                    modifier = Modifier
-                                        .padding(10.dp)
-                                        .background(Color.White)
-                                ) {
-                                    Text(
-                                        text = "To:",
-                                        fontSize = 14.sp,
-                                        modifier = Modifier
-                                            .padding(horizontal = 10.dp)
-                                            .fillMaxWidth()
-                                    )
-                                    for (expenseWithUser in toWhom.filter { it.getAmount() <= 0 }) {
-                                        val user = expenseWithUser.getUser()
-                                        if (user != null) {
-                                            Row(
-                                                modifier = Modifier.fillMaxWidth(),
-                                                horizontalArrangement = Arrangement.SpaceBetween
-                                            ) {
-                                                Text(
-                                                    text = user.getUsername(),
-                                                    fontSize = 14.sp,
-                                                    modifier = Modifier
-                                                        .padding(horizontal = 10.dp)
-                                                )
-                                                Text(
-                                                    text = "${expenseWithUser.getAmount()} $groupCurrency",
-                                                    fontSize = 14.sp,
-                                                    modifier = Modifier
-                                                        .padding(horizontal = 10.dp)
-                                                )
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                            val coroutineScope = rememberCoroutineScope()
-                            val imgBase64 =
-                                BackendService(preferences).getExpensePicture(item.getGlobalId())
-                            val imageBitmap = remember { mutableStateOf<ImageBitmap?>(null) }
-
-                            DisposableEffect(Unit) {
-                                if (imgBase64 != null && imageBitmap.value == null) {
-                                    coroutineScope.launch {
-                                        val bitmap = withContext(Dispatchers.IO) {
-                                            BitmapFactory.decodeByteArray(
-                                                imgBase64,
-                                                0,
-                                                imgBase64.size
-                                            ).asImageBitmap()
-                                        }
-                                        imageBitmap.value = bitmap
-                                    }
-                                }
-                                onDispose { }
-                            }
-
-                            if (imageBitmap.value != null) {
-                                Image(
-                                    bitmap = imageBitmap.value!!,
-                                    contentDescription = "Image",
-                                    modifier = Modifier
-                                        .padding(horizontal = 10.dp)
-                                        .fillMaxWidth()
-                                )
-                            }
-                        }
-                    }
-                }
-            )
-        }
+        GroupSpendingList(
+            data,
+            preferences,
+            groupId,
+            context,
+            refreshAndOpenGroup
+        )
     }
 }
